requirements.txt
Flask==1.1.2
qrcode==6.1
firebase_admin==5.0.1
flask_cors==3.0.10
gunicorn
pillow
python-dateutil